backend/scripts/requirements.txt
attrs>=21.4.0
beautifulsoup4>=4.9.3
bs4>=0.0.1
certifi>=2021.10.8
chromedriver-autoinstaller>=0.3.1
exceptiongroup>=1.1.1
fastapi>=0.95.2
h11>=0.12.0
idna>=3.3
motor>=2.5.1
numpy>=1.20.0,<2.0.0
outcome>=1.1.0
packaging>=21.3
pandas>=1.3.5
pymongo>=3.12.3
PySocks>=1.7.1
python-dateutil>=2.8.2
python-dotenv>=0.19.2
pytz>=2021.3
Scweet>=1.8
selenium>=4.1.0
six>=1.16.0
sniffio>=1.2.0
sortedcontainers>=2.4.0
soupsieve>=2.3.1
trio>=0.19.0
trio-websocket>=0.9.2
typing_extensions>=4.1.1
tzdata>=2021.5
urllib3>=1.26.8
websocket-client>=1.2.3
wsproto>=1.0.0
Flask>=2.0.0